requirements.txt
# Core
anthropic>=0.25.0
streamlit>=1.32.0
python-dotenv>=1.0.0

# ML (existing --- do not change versions)
xgboost>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0

# RAG + Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0

# Nutrition evaluation
datasets>=2.18.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
